Code changes done as suggested in peer review
