+++ DigitalLearningSolutions.Web/Controllers/MyAccountController.cs
@@ -138,24 +138,7 @@
             if (!ModelState.IsValid)
             {
                 var model = new EditDetailsViewModel(formData, dlsSubApplication);
-                var validationResults = model.Validate(new ValidationContext(model, null, null));
-                foreach (var result in validationResults)
-                {
-                    var key = string.Join("", result.MemberNames);
-                    if (ModelState.TryGetValue(key, out var value))
-                    {
-                        if (value.ValidationState == ModelValidationState.Invalid)
-                        {
-                            continue;
-                        }
-                    }
-
-                    ModelState.TryAddModelError(
-                        key,
-                        result.ErrorMessage
-                    );
-                }
-
+                model.ValidateManually(ModelState);
                 return View(model);
                 return ReturnToEditDetailsViewWithErrors(formData, dlsSubApplication, userDelegateId);
             }
